+++ datacontract/catalog/catalog.py
@@ -10,8 +10,10 @@ from datacontract.export.html_export import get_version
 from datacontract.model.data_contract_specification import DataContractSpecification
 
 
-def create_data_contract_html(contracts, file: Path, path: Path):
-    data_contract = DataContract(data_contract_file=f"{file.absolute()}", inline_definitions=True, inline_quality=True)
+def create_data_contract_html(contracts, file: Path, path: Path, schema: str):
+    data_contract = DataContract(
+        data_contract_file=f"{file.absolute()}", inline_definitions=True, inline_quality=True, schema_location=schema
+    )
     html = data_contract.export(export_format="html")
     spec = data_contract.get_data_contract_specification()
     file_without_suffix = file.with_suffix(".html")